# docs/openapi/openapi.json
{
  "components": {
    "schemas": {
      "AnalyzeRequest": {
        "description": "POST /analyze 請求。",
        "properties": {
          "database": {
            "default": "",
            "title": "Database",
            "type": "string"
          },
          "db_name": {
            "default": "",
            "title": "Db Name",
            "type": "string"
          },
          "db_server": {
            "default": "",
            "title": "Db Server",
            "type": "string"
          },
          "expand_depth": {
            "default": 0,
            "title": "Expand Depth",
            "type": "integer"
          },
          "expand_max_programs": {
            "default": 10,
            "title": "Expand Max Programs",
            "type": "integer"
          },
          "include_execution_paths": {
            "default": true,
            "title": "Include Execution Paths",
            "type": "boolean"
          },
          "include_snippets": {
            "default": true,
            "title": "Include Snippets",
            "type": "boolean"
          },
          "include_sp_defs": {
            "default": false,
            "title": "Include Sp Defs",
            "type": "boolean"
          },
          "include_view_layer": {
            "default": false,
            "title": "Include View Layer",
            "type": "boolean"
          },
          "max_paths": {
            "anyOf": [
              {
                "maximum": 60.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Paths"
          },
          "program_names": {
            "items": {
              "type": "string"
            },
            "title": "Program Names",
            "type": "array"
          },
          "question": {
            "default": "",
            "title": "Question",
            "type": "string"
          },
          "refresh": {
            "default": false,
            "title": "Refresh",
            "type": "boolean"
          },
          "source": {
            "$ref": "#/components/schemas/AzureSource"
          },
          "system": {
            "default": "",
            "title": "System",
            "type": "string"
          },
          "wrapper_contract": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": "",
            "title": "Wrapper Contract"
          }
        },
        "required": [
          "program_names"
        ],
        "title": "AnalyzeRequest",
        "type": "object"
      },
      "AnalyzeResponse": {
        "properties": {
          "not_found": {
            "items": {
              "type": "string"
            },
            "title": "Not Found",
            "type": "array"
          },
          "programs": {
            "items": {
              "$ref": "#/components/schemas/ProgramAnalysis"
            },
            "title": "Programs",
            "type": "array"
          },
          "source_root": {
            "default": "",
            "title": "Source Root",
            "type": "string"
          }
        },
        "title": "AnalyzeResponse",
        "type": "object"
      },
...
      "FindBySPRequest": {
        "description": "POST /find_by_sp 請求：反查「哪些程式呼叫了這支 SP」（純快取比對，不觸發 clone）。\n\ncache_only=True（預設）時，若該 repo 尚未 clone 過，直接回傳 skipped=True，\n不會觸發 Azure clone —— 這支端點常被逐系統掃描式呼叫（不知道 SP 屬於哪個\n系統），避免對每個未分析過的系統都觸發一次昂貴的 clone。",
        "properties": {
          "cache_only": {
            "default": true,
            "title": "Cache Only",
            "type": "boolean"
          },
          "database": {
            "default": "",
            "title": "Database",
            "type": "string"
          },
          "refresh": {
            "default": false,
            "title": "Refresh",
            "type": "boolean"
          },
          "source": {
            "$ref": "#/components/schemas/AzureSource"
          },
          "sp_name": {
            "title": "Sp Name",
            "type": "string"
          },
          "wrapper_contract": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": "",
            "title": "Wrapper Contract"
          }
        },
        "required": [
          "sp_name"
        ],
        "title": "FindBySPRequest",
        "type": "object"
      },
      "FindBySPResponse": {
        "properties": {
          "diagnostics": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Diagnostics",
            "type": "array"
          },
          "matches": {
            "items": {
              "$ref": "#/components/schemas/SPMatchProgram"
            },
            "title": "Matches",
            "type": "array"
          },
          "skipped": {
            "default": false,
            "title": "Skipped",
            "type": "boolean"
          },
          "source_root": {
            "default": "",
            "title": "Source Root",
            "type": "string"
          },
          "sp_name": {
            "default": "",
            "title": "Sp Name",
            "type": "string"
          }
        },
        "title": "FindBySPResponse",
        "type": "object"
      },
      "FindByTableRequest": {
        "description": "POST /find_by_table 請求：反查「哪些程式存取了這張資料表」（純快取比對，不觸發 clone）。\n\n情境：使用者打算異動某張資料表（改欄位、改約束等），需要先知道哪些程式會受影響——\n這跟 FindBySPRequest 是同一種「反查」需求；C# inline SQL facts 與 SQL Execution\nGraph 分別提供直接與 SQL module lineage。cache_only=True（預設）時，若該系統\n尚未分析過就直接跳過，不觸發 clone（供逐系統嘗試反查時使用）。",
        "properties": {
          "cache_only": {
            "default": true,
            "title": "Cache Only",
            "type": "boolean"
          },
          "database": {
            "default": "",
            "title": "Database",
            "type": "string"
          },
          "refresh": {
            "default": false,
            "title": "Refresh",
            "type": "boolean"
          },
          "source": {
            "$ref": "#/components/schemas/AzureSource"
          },
          "table_name": {
            "title": "Table Name",
            "type": "string"
          },
          "wrapper_contract": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": "",
            "title": "Wrapper Contract"
          },
          "write_only": {
            "default": false,
            "title": "Write Only",
            "type": "boolean"
          }
        },
        "required": [
          "table_name"
        ],
        "title": "FindByTableRequest",
        "type": "object"
      },
      "FindByTableResponse": {
        "properties": {
          "diagnostics": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Diagnostics",
            "type": "array"
          },
          "excluded_count": {
            "default": 0,
            "title": "Excluded Count",
            "type": "integer"
          },
          "matches": {
            "items": {
              "$ref": "#/components/schemas/TableMatchProgram"
            },
            "title": "Matches",
            "type": "array"
          },
          "skipped": {
            "default": false,
            "title": "Skipped",
            "type": "boolean"
          },
          "source_root": {
            "default": "",
            "title": "Source Root",
            "type": "string"
          },
          "table_name": {
            "default": "",
            "title": "Table Name",
            "type": "string"
          }
        },
        "title": "FindByTableResponse",
        "type": "object"
      },
...
      "FlowChainResponse": {
        "properties": {
          "backward_chains": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Backward Chains",
            "type": "array"
          },
          "diagnostics": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Diagnostics",
            "type": "array"
          },
          "direction": {
            "default": "",
            "title": "Direction",
            "type": "string"
          },
          "forward_chain": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forward Chain"
          },
          "skipped": {
            "default": false,
            "title": "Skipped",
            "type": "boolean"
          },
          "source_root": {
            "default": "",
            "title": "Source Root",
            "type": "string"
          }
        },
        "title": "FlowChainResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "程式碼/DB 靜態影響分析服務（供 spec-rag 協調呼叫，不含 AI）",
    "title": "Impact Analysis Service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/analyze": {
      "post": {
        "operationId": "analyze_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze"
      }
    },
    "/find_by_sp": {
      "post": {
        "description": "反查「哪些程式呼叫了這支 SP」（純快取比對；cache_only=True 時不觸發 clone）。",
        "operationId": "find_by_sp_find_by_sp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindBySPRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindBySPResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Find By Sp"
      }
    },
    "/find_by_table": {
      "post": {
        "description": "反查「哪些程式存取了這張資料表」（純快取比對；cache_only=True 時不觸發 clone）。",
        "operationId": "find_by_table_find_by_table_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindByTableRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindByTableResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Find By Table"
      }
    },
    "/flow_chain": {
      "post": {
        "description": "組出變更影響「關係鏈」候選清單（純靜態組裝，無 AI 判斷；cache_only=True 時不觸發 clone）。\n\ndirection=\"forward\" 時 program_name/anchor_method 必填；\ndirection=\"backward\" 時 table_name 必填（column_name 選填，僅近似文字比對）。",
        "operationId": "flow_chain_flow_chain_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlowChainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowChainResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Flow Chain"
      }
    },
...
    "/locate_object": {
      "post": {
        "description": "從 Object Location Index 猜哪些 Database 可能持有指定物件名稱，不開任何 SQL 快取。",
        "operationId": "locate_object_locate_object_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocateObjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocateObjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Locate Object"
      }
    },
    "/path_evidence": {
      "post": {
        "description": "依 path_id 取得目前 source/SQL snapshot 對應的精確 evidence。",
        "operationId": "path_evidence_path_evidence_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PathEvidenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PathEvidenceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Path Evidence"
      }
    },
    "/refresh": {
      "post": {
        "description": "更新指令：git pull 取得最新程式碼並重新解析、覆寫快取。",
        "operationId": "refresh_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh"
      }
    },
    "/refresh_sql": {
      "post": {
        "description": "更新 SQL 快取指令：重新連線 SQL Server 撈取整庫 SP/View/Function 定義與\n資料表 Schema，使用 ScriptDom 建立 SQL Execution Graph，覆寫本機落地快取\n（data/sql_cache/）。\n\nserver/db_name 由呼叫端（catalog）提供，缺一即報錯、不嘗試連線。連線身分預設\n用 .env 的全域 DB_AUTH_MODE；請求帶了完整的 db_user_id/db_password 才改用那組\n覆寫（ADR-0010）。",
        "operationId": "refresh_sql_refresh_sql_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshSqlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshSqlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Sql"
      }
    },
    "/refresh_sql/status/{job_id}": {
      "get": {
        "description": "查詢 SQL 快取更新的即時進度。",
        "operationId": "refresh_sql_status_refresh_sql_status__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshSqlProgressResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Sql Status"
      }
    },
...
    "/wrapper_contracts/accept": {
      "post": {
        "description": "Preview or explicitly apply one reviewed external-wrapper proposal.",
        "operationId": "accept_wrapper_contract_wrapper_contracts_accept_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WrapperContractAcceptanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrapperContractAcceptanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept Wrapper Contract"
      }
    }
  }
}